Chi - debugged with Dino, had issues when adding data to db, crashed app:
The error you are experiencing is likely related to the fact that you are passing a string value expiryDateString to the set() method of a Firestore document reference. The set() method expects a Map<String, Object> that contains the data to be written to the document.

To fix this error, you need to create a Map object that contains the expiryDateString value and pass it to the set() method. Here's an example of how you could modify your code to do this:

val expiryDateString = recognizedTextEt.text.toString()

val data = hashMapOf(
    "expiryDate" to expiryDateString
)

db.collection("expiryDate").document("new-expiry-date").set(data)

In this modified code, we create a HashMap object called data that contains a single key-value pair: the key is "expiryDate" and the value is the expiryDateString. We then pass the data object to the set() method.

This should resolve the error you are experiencing and allow you to write the expiryDateString value to the Firestore database.